Splitting up `octoshelf.js` to smaller files
My plan is to start writing tests on octoshelf.js. However, the
logic for the core app experience (managing repos and prs) is
mingled with the animations and event handlers.

Because of how tied together everything is, it makes it harder to
write tests.

So I pulled out anything animation related into its own file, and
anything having to do with the actionPanel in a separate file.

I'm planning on writing more tests shortly, this is just prep.
