+++ views/index.ejs
@@ -56,7 +56,7 @@
     <section id="appBackground" class="app-background"></section>
 </main>
 <section class="infoPanel">
-    <section class="infoPanel-topPanel">
+    <section class="infoPanel-actionPanel">
         <h1>OctoShelf</h1><span class="tagline">, <span class="tagline-githubPowered">a
             <a href="https://github.com" class="octicon octicon-logo-github"></a>
             powered</span> pull request manager
